scripts: Move generators to seperate folder
