+++ scripts/generators/best_practices_generator.py
