+++ scripts/generators/command_validation_generator.py
