+++ scripts/generators/external_revision_generator.py
