+++ scripts/generators/object_tracker_generator.py
