+++ scripts/generators/thread_safety_generator.py
